ant-d config

+++ src/api/request.js
@@ -1,5 +1,6 @@
 import axios from "axios";
 import { BASE_URL } from "../constants/api";
+import { isClient } from "../helpers/isClient";
 
 const request = axios.create({
   baseURL: BASE_URL,
@@ -8,7 +9,7 @@ const request = axios.create({
 
 request.interceptors.request.use(
   (config) => {
-    const user = JSON.parse(localStorage.getItem("user"));
+    const user = JSON.parse(isClient ? localStorage.getItem("user") : {});
 
     config.headers["Authorization"] = `Bearer ${user?.access_token}`;
 
@@ -25,8 +26,8 @@ request.interceptors.response.use(
   },
   (error) => {
     if (error?.response?.status === 401) {
-      window.location.assign("/login");
-      localStorage.clear();
+      isClient && window.location.assign("/login");
+      isClient && localStorage.clear();
     }
     return error.response;
   }
